+++ rocketpy/__init__.py
@@ -46,8 +46,9 @@ from .rocket import (
 )
 from .sensitivity import SensitivityModel
 from .sensors import Accelerometer, Barometer, GnssReceiver, Gyroscope
-from .simulation import Flight, MonteCarlo
+from .simulation import Flight, MonteCarlo, MultivariateRejectionSampler
 from .stochastic import (
+    StochasticAirBrakes,
     StochasticEllipticalFins,
     StochasticEnvironment,
     StochasticFlight,