+++ rocketpy/plots/rocket_plots.py
@@ -1,5 +1,3 @@
-import warnings
-
 import matplotlib.pyplot as plt
 import numpy as np
 
@@ -88,40 +86,6 @@ class _RocketPlots:
             disp_type="surface",
             alpha=1,
         )
-
-    def power_on_drag(self):
-        """Plots power on drag of the rocket as a function of time.
-
-        Returns
-        -------
-        None
-        """
-
-        warnings.warn(
-            "The method 'power_on_drag' is deprecated as of version "
-            + "1.2 and will be removed in version 1.4 "
-            + "Use 'plots.drag_curves' instead.",
-            DeprecationWarning,
-        )
-
-        self.rocket.power_on_drag()
-
-    def power_off_drag(self):
-        """Plots power off drag of the rocket as a function of time.
-
-        Returns
-        -------
-        None
-        """
-
-        warnings.warn(
-            "The method 'power_off_drag' is deprecated as of version "
-            + "1.2 and will be removed in version 1.4 "
-            + "Use 'plots.drag_curves' instead.",
-            DeprecationWarning,
-        )
-
-        self.rocket.power_off_drag()
 
     # pylint: disable=too-many-statements
     def drag_curves(self, *, filename=None):