+++ rocketpy/stochastic/stochastic_rocket.py
@@ -3,11 +3,13 @@
 import warnings
 from random import choice
 
+from rocketpy.control import _Controller
 from rocketpy.mathutils.vector_matrix import Vector
 from rocketpy.motors.empty_motor import EmptyMotor
 from rocketpy.motors.motor import GenericMotor, Motor
 from rocketpy.motors.solid_motor import SolidMotor
 from rocketpy.rocket.aero_surface import (
+    AirBrakes,
     EllipticalFins,
     NoseCone,
     RailButtons,
@@ -21,6 +23,7 @@ from rocketpy.stochastic.stochastic_generic_motor import StochasticGenericMotor
 from rocketpy.stochastic.stochastic_motor_model import StochasticMotorModel
 
 from .stochastic_aero_surfaces import (
+    StochasticAirBrakes,
     StochasticEllipticalFins,
     StochasticNoseCone,
     StochasticRailButtons,
@@ -148,6 +151,7 @@ class StochasticRocket(StochasticModel):
         self.motors = Components()
         self.aerodynamic_surfaces = Components()
         self.rail_buttons = Components()
+        self.air_brakes = []
         self.parachutes = []
         self.__components_map = {}
         super().__init__(
@@ -389,6 +393,125 @@ class StochasticRocket(StochasticModel):
             rail_buttons, self._validate_position(rail_buttons, lower_button_position)
         )
 
+    def add_air_brakes(self, air_brakes, controller):
+        """Adds an air brake to the stochastic rocket.
+
+        Parameters
+        ----------
+        air_brakes : StochasticAirBrakes or Airbrakes
+            The air brake to be added to the stochastic rocket.
+        controller : _Controller
+            Deterministic air brake controller.
+        """
+        if not isinstance(air_brakes, (AirBrakes, StochasticAirBrakes)):
+            raise TypeError(
+                "`air_brake` must be of AirBrakes or StochasticAirBrakes type"
+            )
+        if isinstance(air_brakes, AirBrakes):
+            air_brakes = StochasticAirBrakes(air_brakes=air_brakes)
+
+        self.air_brakes.append(air_brakes)
+        self.air_brake_controller = controller
+
+    def add_cp_eccentricity(self, x=None, y=None):
+        """Moves line of action of aerodynamic forces to simulate an
+        eccentricity in the position of the center of pressure relative
+        to the center of dry mass of the rocket.
+
+        Parameters
+        ----------
+        x : tuple, list, int, float, optional
+            Distance in meters by which the CP is to be translated in
+            the x direction relative to the center of dry mass axial line.
+            The x axis is defined according to the body axes coordinate system.
+        y : tuple, list, int, float, optional
+            Distance in meters by which the CP is to be translated in
+            the y direction relative to the center of dry mass axial line.
+            The y axis is defined according to the body axes coordinate system.
+
+        Returns
+        -------
+        self : StochasticRocket
+            Object of the StochasticRocket class.
+        """
+        self.cp_eccentricity_x = self._validate_eccentricity("cp_eccentricity_x", x)
+        self.cp_eccentricity_y = self._validate_eccentricity("cp_eccentricity_y", y)
+        return self
+
+    def add_thrust_eccentricity(self, x=None, y=None):
+        """Moves line of action of thrust forces to simulate a
+        misalignment of the thrust vector and the center of dry mass.
+
+        Parameters
+        ----------
+        x : tuple, list, int, float, optional
+            Distance in meters by which the line of action of the
+            thrust force is to be translated in the x direction
+            relative to the center of dry mass axial line. The x axis
+            is defined according to the body axes coordinate system.
+        y : tuple, list, int, float, optional
+            Distance in meters by which the line of action of the
+            thrust force is to be translated in the y direction
+            relative to the center of dry mass axial line. The y axis
+            is defined according to the body axes coordinate system.
+
+        Returns
+        -------
+        self : StochasticRocket
+            Object of the StochasticRocket class.
+        """
+        self.thrust_eccentricity_x = self._validate_eccentricity(
+            "thrust_eccentricity_x", x
+        )
+        self.thrust_eccentricity_y = self._validate_eccentricity(
+            "thrust_eccentricity_y", y
+        )
+        return self
+
+    def _validate_eccentricity(self, eccentricity, position):
+        """Validate the eccentricity argument.
+
+        Parameters
+        ----------
+        eccentricity : str
+            The eccentricity to which the position argument refers to.
+        position : tuple, list, int, float
+            The position argument to be validated.
+
+        Returns
+        -------
+        tuple or list
+            Validated position argument.
+
+        Raises
+        ------
+        ValueError
+            If the position argument does not conform to the specified formats.
+        """
+        if isinstance(position, tuple):
+            return self._validate_tuple(
+                eccentricity,
+                position,
+            )
+        elif isinstance(position, (int, float)):
+            return self._validate_scalar(
+                eccentricity,
+                position,
+            )
+        elif isinstance(position, list):
+            return self._validate_list(
+                eccentricity,
+                position,
+            )
+        elif position is None:
+            position = []
+            return self._validate_list(
+                eccentricity,
+                position,
+            )
+        else:
+            raise AssertionError("`position` must be a tuple, list, int, or float")
+
     def _validate_position(self, validated_object, position):
         """Validate the position argument.
 
@@ -531,6 +654,7 @@ class StochasticRocket(StochasticModel):
         generated_dict["motors"] = []
         generated_dict["aerodynamic_surfaces"] = []
         generated_dict["rail_buttons"] = []
+        generated_dict["air_brakes"] = []
         generated_dict["parachutes"] = []
         self.last_rnd_dict = generated_dict
         yield generated_dict
@@ -573,10 +697,22 @@ class StochasticRocket(StochasticModel):
         )
         return rail_buttons, lower_button_position_rnd, upper_button_position_rnd
 
+    def _create_air_brake(self, stochastic_air_brake):
+        air_brake = stochastic_air_brake.create_object()
+        self.last_rnd_dict["air_brakes"].append(stochastic_air_brake.last_rnd_dict)
+        return air_brake
+
     def _create_parachute(self, stochastic_parachute):
         parachute = stochastic_parachute.create_object()
         self.last_rnd_dict["parachutes"].append(stochastic_parachute.last_rnd_dict)
         return parachute
+
+    def _create_eccentricities(self, stochastic_x, stochastic_y, eccentricity):
+        x_rnd = self._randomize_position(stochastic_x)
+        self.last_rnd_dict[eccentricity + "_x"] = x_rnd
+        y_rnd = self._randomize_position(stochastic_y)
+        self.last_rnd_dict[eccentricity + "_y"] = y_rnd
+        return x_rnd, y_rnd
 
     def create_object(self):
         """Creates and returns a Rocket object from the randomly generated input
@@ -609,6 +745,23 @@ class StochasticRocket(StochasticModel):
         rocket.power_off_drag *= generated_dict["power_off_drag_factor"]
         rocket.power_on_drag *= generated_dict["power_on_drag_factor"]
 
+        if hasattr(self, "cp_eccentricity_x") and hasattr(self, "cp_eccentricity_y"):
+            cp_ecc_x, cp_ecc_y = self._create_eccentricities(
+                self.cp_eccentricity_x,
+                self.cp_eccentricity_y,
+                "cp_eccentricity",
+            )
+            rocket.add_cp_eccentricity(cp_ecc_x, cp_ecc_y)
+        if hasattr(self, "thrust_eccentricity_x") and hasattr(
+            self, "thrust_eccentricity_y"
+        ):
+            thrust_ecc_x, thrust_ecc_y = self._create_eccentricities(
+                self.thrust_eccentricity_x,
+                self.thrust_eccentricity_y,
+                "thrust_eccentricity",
+            )
+            rocket.add_thrust_eccentricity(thrust_ecc_x, thrust_ecc_y)
+
         for component_motor in self.motors:
             motor, position_rnd = self._create_motor(component_motor)
             rocket.add_motor(motor, position_rnd)
@@ -616,6 +769,17 @@ class StochasticRocket(StochasticModel):
         for component_surface in self.aerodynamic_surfaces:
             surface, position_rnd = self._create_surface(component_surface)
             rocket.add_surfaces(surface, position_rnd)
+
+        for air_brake in self.air_brakes:
+            air_brake = self._create_air_brake(air_brake)
+            _controller = _Controller(
+                interactive_objects=air_brake,
+                controller_function=self.air_brake_controller.base_controller_function,
+                sampling_rate=self.air_brake_controller.sampling_rate,
+                initial_observed_variables=self.air_brake_controller.initial_observed_variables,
+            )
+            rocket.air_brakes.append(air_brake)
+            rocket._add_controllers(_controller)
 
         for component_rail_buttons in self.rail_buttons:
             (